restructure data tables as their own components
