+++ app/(main)/_components/ParticipantsTable/Columns.tsx
@@ -1,0 +1,58 @@
+'use client';
+
+import { type ColumnDef } from '@tanstack/react-table';
+import type { User } from '@prisma/client';
+import { ArrowUpDown } from 'lucide-react';
+import { Button } from '~/components/ui/Button';
+import { ActionsDropdown } from '~/components/DataTable/ActionsDropdown';
+import { Checkbox } from '~/components/ui/checkbox';
+
+export const ParticipantColumns: ColumnDef<User>[] = [
+  {
+    id: 'select',
+    header: ({ table }) => (
+      <Checkbox
+        checked={table.getIsAllPageRowsSelected()}
+        onCheckedChange={(value) => table.toggleAllPageRowsSelected(!!value)}
+        aria-label="Select all"
+      />
+    ),
+    cell: ({ row }) => (
+      <Checkbox
+        checked={row.getIsSelected()}
+        onCheckedChange={(value) => row.toggleSelected(!!value)}
+        aria-label="Select row"
+      />
+    ),
+    enableSorting: false,
+    enableHiding: false,
+  },
+  {
+    accessorKey: 'id',
+    header: ({ column }) => {
+      return (
+        <Button
+          variant="ghost"
+          onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
+        >
+          Participant ID
+          <ArrowUpDown className="ml-2 h-4 w-4" />
+        </Button>
+      );
+    },
+  },
+  {
+    accessorKey: 'name',
+    header: 'Name',
+  },
+  {
+    accessorKey: 'email',
+    header: 'E-mail',
+  },
+  {
+    id: 'actions',
+    cell: () => {
+      return <ActionsDropdown menuItems={['Edit', 'Delete']} />;
+    },
+  },
+];
